Fixed a bug in travis parse-build.
Converting the old :matrix field to the new :job-ids field was done in a order
dependant way. Sometimes when no :matrix field was present, update-in combined
with map created an empty seq as a value for :job-ids.

Moved the replacement of underscores with hyphens to a parse-json function.
Rewrote the parse-build helper functions to be order independent.